+++ rspec-expectations/lib/rspec/matchers/built_in/change.rb
@@ -114,12 +114,10 @@ module RSpec
         end
 
         def positive_failure_reason
-          return "was not given a block" unless Proc === @event_proc
           "is still #{@actual_before_description}"
         end
 
         def negative_failure_reason
-          return "was not given a block" unless Proc === @event_proc
           "did change from #{@actual_before_description} " \
             "to #{description_of change_details.actual_after}"
         end
@@ -173,7 +171,6 @@ module RSpec
       private
 
         def failure_reason
-          return "was not given a block" unless Proc === @event_proc
           "was changed by #{description_of @change_details.actual_delta}"
         end
       end
@@ -202,7 +199,6 @@ module RSpec
 
         # @private
         def failure_message
-          return not_given_a_block_failure unless Proc === @event_proc
           return before_value_failure      unless @matches_before
           return did_not_change_failure    unless @change_details.changed?
           after_value_failure
@@ -259,11 +255,6 @@ module RSpec
             "did change from #{@actual_before_description} " \
             "to #{description_of @change_details.actual_after}"
         end
-
-        def not_given_a_block_failure
-          "expected #{@change_details.value_representation} to have changed " \
-            "#{change_description}, but was not given a block"
-        end
       end
 
       # @api private
@@ -295,7 +286,6 @@ module RSpec
 
         # @private
         def failure_message_when_negated
-          return not_given_a_block_failure unless Proc === @event_proc
           return before_value_failure unless @matches_before
           did_change_failure
         end