+++ rspec-expectations/Changelog.md
@@ -10,6 +10,8 @@ Breaking Changes:
 * Remove `include_chain_clauses_in_custom_matcher_descriptions` option
   and make it the default. (Phil Pirozhkov, rspec/rspec-expectations#1279)
 * Remove support for present-tense dynamic predicate. (Phil Pirozhkov, rspec/rspec-expectations#1286)
+* Prevent implicit blocks (e.g blocks as values) from being used with block matchers.
+  (Phil Pirozhkov, rspec/rspec-expectations#1285)
 
 Enhancements: